+++ com/logout/http/handlers/prompt.js
@@ -7,9 +7,9 @@ exports = module.exports = function(csrfProtection, state) {
     res.locals.csrfToken = req.csrfToken();
     
     // NOTE: This will include locals for state.
-    res.render('prompt', function(err, str) {
+    res.render('logout', function(err, str) {
       if (err && err.view) {
-        var view = path.resolve(__dirname, '../../views/logout.ejs');
+        var view = path.resolve(__dirname, '../views/prompt.ejs');
         ejs.renderFile(view, res.locals, function(err, str) {
           if (err) { return next(err); }
           res.send(str);