+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `/signup` endpoint.  Moved from [`@authnomicon/account`](https://github.com/authnomicon/account).
 - Added `/account/select` endpoint.  Moved from [`@authnomicon/account`](https://github.com/authnomicon/account).
 - Added login prompt.  Moved from [`@authnomicon/prompts`](https://github.com/authnomicon/prompts).
+- Added logout prompt.  Moved from [`@authnomicon/prompts`](https://github.com/authnomicon/prompts).
 - Added select-account prompt.  Moved from [`@authnomicon/prompts`](https://github.com/authnomicon/prompts).
 
 ### Changed